fix forceNew on vms from linked clones on apply
When a VM is cloned as a linked clone, every time you ran terraform
apply, it would cause a force_new and destroy and readd the vm.

Since the API does not have a way to indicate if the device was a linked
clone or full clone, this was hardcoded.

Due to the way the `GetOk(key)` function works, this would always return
false, which would then cause the full_clone value to be reset back to
it's default, which is true.

Fixes by moving this to be a statically set value, since full_clone has
a default, which means it will always be set and we can force this in
the schema/state.