Support PXE for install usecase
Currently, the PXE usecase does not allow for using PXE for provisioning and booting from a disk, as the PXE mode requires network to be the first option.

By setting the boot order to the hard drive first and then network will allow hosts to be provisioned on the first boot and run from disk on subsequent boots.

Signed-off-by: Jacob McSwain <jacob@mcswain.dev>