Fix parallel execution handling when returning early (#930)
resourceVmQemuCreate and resourceVmQemuUpdate can both return early due to
errors. When this happens, we must ensure that we "unlock" the slot we have
claimed for parallel execution. This is safe to do in a defer action because
even if we manually "unlock" prior to the function return, the unlock action
checks to ensure that the "lock" is held prior to attempting to return a slot
to the pool.